+++ lesson/doc/不带鬼牌的查表胡牌算法.txt
@@ -1,0 +1,40 @@
+mjlib胡牌思路 注意：不带鬼牌
+
+代码路径：https://github.com/yuanfengyun/qipai/tree/master/mjlib_lua/base_table
+
+1、不同花色分开处理，分别校验是否能满足 将、顺子、刻子
+2、分析东南西北风中发白时，不需要分析顺子的情况，简单很多
+3、分析单一花色时，按连续性将麻将切割成小段
+比如：1筒2筒3筒3筒3筒3筒6筒7筒8筒2万3万3万3万4万
+拆成：a段 1筒2筒3筒3筒3筒3筒
+      b段 6筒7筒8筒
+      c段 2万3万3万3万4万
+
+4、将每小段去掉花色信息、点数信息以后只剩下形状
+a段变成 114
+b段变成 111
+c段变成 131
+
+5、如果小段中，牌的数量为 3*n
+查表wave_tbl判断是否满足拆分, 例，a的key为114, b的key为111
+查不到表示不满足拆分，不能胡
+
+6、如果小段中，牌的数量为 3*n + 2，代表里面有将牌
+查表wave_tbl_eye判断是否满足带将的拆分，例，c的key为131
+查不到表示不满足拆分，不能胡
+
+7、组合表例子:
+wave_tbl = {
+    [111] = true,[222]=true,[333]=true,[444]=true,
+    [114] = true, [141]=true, [411]=true,
+    [144] = true, [414]=true, [441]=true,
+}
+
+8、带将的组合表例子：
+wave_eye_tbl = {[2] = true, [2111] = true}
+
+9、表生成：
+  代码：https://github.com/yuanfengyun/qipai/blob/master/mjlib_lua/base_table/gen_table.lua
+  思路：
+      穷举单种花色所有不带将组合情况（依次放次1组、2组、3组、4组顺子或刻子，将最到得到的牌按连续性切分放入wave表中）
+      穷举单种花色所有带将组合情况（先放入一对将，然后依次放次1组、2组、3组、4组顺子或刻子，将最到得到的牌按连续性切分放入wave表或wave_eye表中）
