+++ lesson/doc/网狐算法思路.txt
@@ -1,0 +1,12 @@
+2A + n*ABC + n*AAA
+
+1万1万1万2万2万3万4万5万
+
+拆解出所有可能性
+1万1万1万
+1万2万3万
+2万3万4万
+3万4万5万
+
+排列组合选四组
+C 4 n
