+++ lesson/doc/麻将回溯算法.txt
@@ -1,0 +1,17 @@
+麻将回溯算法
+
+麻将有1-4张牌
+
+1、一张牌的时候
+   a、拆顺子{1,1,1}
+2、二张牌的时候
+   a、拆1对将 {2,0,0}
+   b、拆两个顺子{2,2,2}
+3、三张牌的时候
+   a、拆1个刻子{3,0,0}
+   b、1对将，1个顺子{3,1,1}
+   c、3个顺子，和a是等价的
+4、四张牌的时候
+   a、拆1个刻子，1个顺子{4,1,1}
+   b、拆1对将，2个顺子{4,2,2}
+   c、4个顺子，和a是等价的
