+++ lesson/mj_huisu/huisu.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool can_hu_huisu(char* cards);
