+++ lesson/mj_lesson.sln
@@ -1,0 +1,59 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lesson1", "lesson1", "{F35EED68-048A-4981-A4A9-8BAB49052E57}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lesson2", "lesson2", "{2591D621-8AB2-4A17-A828-46FEED586018}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lesson3", "lesson3", "{C17BE542-8297-4CBB-A6DA-EC2D88C4C9DB}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mj_huisu", "mj_huisu\mj_lesson.vcxproj", "{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "split", "split\split.vcxproj", "{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "table", "table\table.vcxproj", "{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x64.ActiveCfg = Debug|x64
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x64.Build.0 = Debug|x64
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x86.ActiveCfg = Debug|Win32
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x86.Build.0 = Debug|Win32
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x64.ActiveCfg = Release|x64
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x64.Build.0 = Release|x64
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x86.ActiveCfg = Release|Win32
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x86.Build.0 = Release|Win32
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Debug|x64.ActiveCfg = Debug|x64
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Debug|x64.Build.0 = Debug|x64
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Debug|x86.ActiveCfg = Debug|Win32
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Debug|x86.Build.0 = Debug|Win32
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Release|x64.ActiveCfg = Release|x64
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Release|x64.Build.0 = Release|x64
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Release|x86.ActiveCfg = Release|Win32
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7}.Release|x86.Build.0 = Release|Win32
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Debug|x64.ActiveCfg = Debug|x64
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Debug|x64.Build.0 = Debug|x64
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Debug|x86.ActiveCfg = Debug|Win32
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Debug|x86.Build.0 = Debug|Win32
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Release|x64.ActiveCfg = Release|x64
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Release|x64.Build.0 = Release|x64
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Release|x86.ActiveCfg = Release|Win32
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77} = {F35EED68-048A-4981-A4A9-8BAB49052E57}
+		{31604320-E2CA-41FD-8D92-4CC98D1BBBB7} = {2591D621-8AB2-4A17-A828-46FEED586018}
+		{E6F2308A-7CA9-45BA-9046-3923BBC91FD1} = {2591D621-8AB2-4A17-A828-46FEED586018}
+	EndGlobalSection
+EndGlobal
