+++ lesson/mj_paileizuhe/Release/stdafx.nativecodeanalysis.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<DEFECTS></DEFECTS>
