+++ lesson/split/hulib.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool can_hu(char* cards);
