+++ lesson/table/hulib.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void init_table();
+
+bool can_hu(char* cards);
