--- mj_lesson/mj_lesson.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mj_lesson", "mj_lesson\mj_lesson.vcxproj", "{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x64.ActiveCfg = Debug|x64
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x64.Build.0 = Debug|x64
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x86.ActiveCfg = Debug|Win32
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Debug|x86.Build.0 = Debug|Win32
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x64.ActiveCfg = Release|x64
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x64.Build.0 = Release|x64
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x86.ActiveCfg = Release|Win32
-		{216AE08B-5B6C-4D38-B0F3-CB187CA7AF77}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
