--- mj_lesson/mj_lesson/huisu.h
@@ -1,3 +1,0 @@
-#pragma once
-
-bool can_hu_huisu(char* cards);
