--- mj_lesson/mj_lesson/targetver.h
@@ -1,8 +1,0 @@
-#pragma once
-
-// 包括 SDKDDKVer.h 将定义可用的最高版本的 Windows 平台。
-
-// 如果要为以前的 Windows 平台生成应用程序，请包括 WinSDKVer.h，并将
-// 将 _WIN32_WINNT 宏设置为要支持的平台，然后再包括 SDKDDKVer.h。
-
-#include <SDKDDKVer.h>
